refactor(HGrids): Move templates to shared feature helper, fix some template features.
