+++ templates/jquery/js/grid/grid-export/files/index.html
@@ -53,6 +53,7 @@
     </head>
     <body>
         <div class="group-fields">
+            Export with features: <br />
             <div class="item">
                 <div id="sorting" class="optionContainer"></div>
                 <label class="item-label" for="sorting">Sorting</label>
