+++ templates/jquery/js/hierarchical-grid/hierarchical-grid-custom/files/index.html
@@ -33,9 +33,7 @@
 				primaryKey: "EmployeeID",
 				childrenDataProperty: "Orders",
 				autoGenerateLayouts: false,
-				 features: [
-					 $(Gridfeatures)
-					],
+				features: $(gridfeatures),
 				columnLayouts: [
 					{
 						key: "Orders",
