+++ templates/jquery/js/hierarchical-grid/hierarchical-grid-export/files/index.html
@@ -50,10 +50,8 @@
                    { key: "Country", headerText: "Country", dataType: "string", width: "10%" }
                 ],
                 childrenDataProperty: "Orders",
-				autoGenerateLayouts: false,
-				features: [
-					$(Gridfeatures)
-				],
+                autoGenerateLayouts: false,
+                features: $(gridfeatures),
                 columnLayouts: [
                     {
                         key: "Orders",
