+++ templates/jquery/js/hierarchical-grid/hierarchical-grid/index.ts
@@ -1,12 +1,12 @@
 import * as path from "path";
 import * as fs from "fs-extra";
-import { GridHelper } from "../gridtemplatehelper";
+import { GridHelper } from "../../../../../lib/project-utility/GridHelper";
 import { jQueryTemplate } from "../../../../../lib/templates/jQueryTemplate";
 import { ProjectConfig } from "../../../../../lib/ProjectConfig";
 import { Util } from "../../../../../lib/Util";
 
 class HierarchicalGridTemplate extends jQueryTemplate {
-
+	private gridHelper: GridHelper;
 	extraConfigurations: ControlExtraConfiguration[];
 
 	public userExtraConfiguration: {};
@@ -23,6 +23,10 @@ class HierarchicalGridTemplate extends jQueryTemplate {
 		this.dependencies = ["igHierarchicalGrid"];
 		this.id = "hierarchical-grid";
 		this.hasExtraConfiguration = true;
+
+		this.gridHelper = new GridHelper();
+		this.gridHelper.hierarchical = true;
+		this.gridHelper.space = "    ";
 		var featureConfiguration: ControlExtraConfiguration = {
 			key: "features",
 			choices: ["Sorting", "Paging", "Filtering"],
@@ -36,22 +40,35 @@ class HierarchicalGridTemplate extends jQueryTemplate {
 		this.userExtraConfiguration = extraConfigKeys;
 	}
 	generateFiles(projectPath: string, name: string, ...options: any[]): Promise<boolean> {
-		
 		var success = true,
 			destinationPath = path.join(projectPath, this.folderName(name));
 		//read html
-		var config = {}, features: string;
-		if (this.userExtraConfiguration["features"] !== undefined) {
-			features = "," + GridHelper.generateFeatures(this.userExtraConfiguration["features"]);
-		} else {
-			features = "";
-		}
+		const config = {};
+		this.gridHelper.addFeature("Responsive", {
+			inherit: false,
+			enableVerticalRendering: false,
+			columnSettings: [
+				{
+					columnKey: "Title",
+					classes: "ui-hidden-phone"
+				},
+				{
+					columnKey: "Region",
+					classes: "ui-hidden-phone"
+				},
+				{
+					columnKey: "City",
+					classes: "ui-hidden-phone"
+				}
+			]
+		});
+		const features = this.gridHelper.generateFeatures(this.userExtraConfiguration["features"], 4);
 
-		config["$(Gridfeatures)"] = features;
+		config["$(gridfeatures)"] = features;
 		config["$(componentName)"] = name;
 		config["$(cssBlock)"] = this.getCssTags();
 		config["$(scriptBlock)"] = this.getScriptTags();
-		var pathsConfig = {};
+		const pathsConfig = {};
 		return Util.processTemplates(path.join(__dirname, "files"), destinationPath, config, pathsConfig);
 	}
 	getExtraConfiguration(): ControlExtraConfiguration[] {
